+++ packages/test-base/src/androidAndroidTest/kotlin/io/realm/kotlin/test/shared/notifications/RealmNotificationsTests.kt
@@ -201,9 +201,11 @@ class RealmNotificationsTests : FlowableTests {
     @Test
     fun closingRealmCompletesFlow() {
         runBlocking {
+            val mutex = Mutex(true)
             val observer = async {
-                realm.asFlow().collect { }
+                realm.asFlow().collect { mutex.unlock() }
             }
+            mutex.lock()
             realm.close()
             withTimeout(5.seconds) {
                 observer.await()